UPDATE: New CSS file.
Adding a CSS file for the Odin Project. Took the original HTML project called recipes from the previous chapter called HTML and now for the CSS chapter of the project. The student is asked to take the previous project and now add style to the web page by including CSS elements to add style properties to improve the look of the project.
NOTE: having difficulties in using Git/GitHub. Difficult learning curve. Hopefully project was updated correctly versus duplicated.